Updated credit var from int to float

+++ models/user.py
@@ -1,4 +1,4 @@
 import pydantic
 class User(pydantic.BaseModel):
     username: str
-    credit: int
+    credit: float
